+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Change MSRV to 1.61.
+- Rate-limit requests to `/api/v1/accounts/search` when `resolve` parameter is used by unauthenticated user.
 
 ## [2.8.0] - 2024-01-29